Split lib/common/avb.c/h into separate libraries
This allows pulling in gPTP client support without pulling in AVTP code
and especially without pulling in a dependency on the igb library. No
source had to change since avb.h just includes the broken-out headers,
and there is a CPP option to include the avb_igb.h file.

Makefiles were updated as well, including for all the examples that use
the common avb code.
