+++ examples/common/Makefile
@@ -1,14 +1,16 @@
+DAEMONS_DIR = ../../daemons
+
 CC ?= gcc
 OPT = -O2 -g
-CFLAGS = $(OPT) -Wall -Wextra -Wno-parentheses
+WARN = -Wall -Wextra -Wno-parentheses
+CFLAGS = $(OPT) $(WARN)
+CPPFLAGS = -I$(DAEMONS_DIR)/mrpd -I$(DAEMONS_DIR)/common
 
 all: talker_mrp_client.o listener_mrp_client.o
 
 talker_mrp_client.o: talker_mrp_client.c talker_mrp_client.h
-	$(CC) $(CFLAGS) -I../../daemons/mrpd  -I../../daemons/common -c talker_mrp_client.c
 
 listener_mrp_client.o: listener_mrp_client.c listener_mrp_client.h
-	$(CC) $(CFLAGS) -I../../daemons/mrpd  -I../../daemons/common -c listener_mrp_client.c
 
 clean:
 	$(RM)  talker_mrp_client.o listener_mrp_client.o
