+++ examples/jackd-listener/Makefile
@@ -1,18 +1,24 @@
+MRPCLIENT_DIR = ../common
+MRPLISTENER_OBJS = listener_mrp_client.o
+MRPLISTENER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPLISTENER_OBJS))
+
+DAEMONS_DIR = ../../daemons
+
 CC ?= gcc
 OPT = -O2 -g
-CFLAGS = $(OPT) -Wall -Wextra -Wno-parentheses -std=gnu99
-INCFLAGS = -I../../daemons/mrpd -I../common -I../../daemons/common
+WARN = -Wall -Wextra -Wno-parentheses
+CFLAGS = $(OPT) $(WARN) -std=gnu99
+CPPFLAGS = -I$(DAEMONS_DIR)/mrpd -I$(MRPCLIENT_DIR) -I$(DAEMONS_DIR)/common
 LDLIBS = -lpcap -lsndfile -ljack -lpthread
 
 all: jack_listener
 
-jack_listener: jack_listener.o ../common/listener_mrp_client.o
+jack_listener: jack_listener.o $(MRPLISTENER_TARGETS)
 
 jack_listener.o: jack_listener.c
-	$(CC) $(CFLAGS) $(INCFLAGS) -c jack_listener.c
 
-../common/listener_mrp_client.o:
-	make -C ../common/ listener_mrp_client.o
+$(MRPCLIENT_DIR)/%.o: $(MRPCLIENT_DIR)/%.c $(MRPCLIENT_DIR)/%.h
+	make -C $(MRPCLIENT_DIR) $@
 
 %: %.o
 	$(CC) $^ $(LDLIBS) -o $@
