+++ examples/jackd-talker/Makefile
@@ -1,25 +1,37 @@
+AVBLIB_DIR = ../../lib/common
+AVBLIB_OBJS = avb_avtp.o avb_gptp.o avb_igb.o
+AVBLIB_TARGETS = $(addprefix $(AVBLIB_DIR)/,$(AVBLIB_OBJS))
+
+MRPCLIENT_DIR = ../common
+MRPTALKER_OBJS = talker_mrp_client.o
+MRPTALKER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPTALKER_OBJS))
+
+IGBLIB_DIR = ../../lib/igb
+DAEMONS_DIR = ../../daemons
+
 CC ?= gcc
 OPT = -O2 -g
-CFLAGS = $(OPT) $(INCFLAGS) -Wall -Wextra -Wno-parentheses -std=gnu99
-INCFLAGS = -I../../lib/igb -I../../daemons/mrpd -I../common -I../../lib/common -I../../daemons/common
+WARN = -Wall -Wextra -Wno-parentheses
+CFLAGS = $(OPT) $(WARN) -std=gnu99
+CPPFLAGS = -I$(IGBLIB_DIR) -I$(DAEMONS_DIR)/mrpd -I$(MRPCLIENT_DIR) -I$(AVBLIB_DIR) -I$(DAEMONS_DIR)/common
 LDLIBS = -ligb -lpci -lrt -pthread -ljack
-LDFLAGS = -L../../lib/igb
+LDFLAGS = -L$(IGBLIB_DIR)
+
+.PHONY: all clean
 
 all: jackd_talker
 
-jackd_talker: jackd_talker.o jack.o ../common/talker_mrp_client.o ../../lib/common/avb.o
+jackd_talker: jackd_talker.o jack.o $(MRPTALKER_TARGETS) $(AVBLIB_TARGETS)
 
 jack.o: jack.c jack.h defines.h
-	$(CC) $(CFLAGS) -c jack.c
 
 jackd_talker.o: jackd_talker.c defines.h jack.h
-	$(CC) -c $(INCFLAGS) -I../../daemons/mrpd $(CFLAGS) jackd_talker.c
 
-../../lib/common/avb.o: ../../lib/common/avb.c ../../lib/common/avb.h
-	make -C ../../lib/common/ avb.o
+$(AVBLIB_DIR)/%.o: $(AVBLIB_DIR)/%.h $(AVBLIB_DIR)/%.c
+	make -C $(AVBLIB_DIR) $@
 
-../common/talker_mrp_client.o:
-	make -C ../common/ talker_mrp_client.o
+$(MRPCLIENT_DIR)/%.o: $(MRPCLIENT_DIR)/%.c $(MRPCLIENT_DIR)/%.h
+	make -C $(MRPCLIENT_DIR) $@
 
 %: %.o
 	$(CC) $(LDFLAGS) $^ $(LDLIBS) -o $@
