+++ examples/live_stream/Makefile
@@ -1,30 +1,41 @@
+AVBLIB_DIR = ../../lib/common
+AVBLIB_OBJS = avb_avtp.o avb_gptp.o avb_igb.o
+AVBLIB_TARGETS = $(addprefix $(AVBLIB_DIR)/,$(AVBLIB_OBJS))
+
+MRPCLIENT_DIR = ../common
+MRPTALKER_OBJS = talker_mrp_client.o
+MRPTALKER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPTALKER_OBJS))
+MRPLISTENER_OBJS = listener_mrp_client.o
+MRPLISTENER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPLISTENER_OBJS))
+
+IGBLIB_DIR = ../../lib/igb
+DAEMONS_DIR = ../../daemons
+
 CC?=gcc
 OPT=-O2 -g
-CFLAGS=$(OPT) -Wall -Wextra -Wno-parentheses
-INCFLAGS=-I ../../lib/igb/ -I../../daemons/mrpd -I../common -I../../lib/common -I../../daemons/common
+WARN=-Wall -Wextra -Wno-parentheses
+CFLAGS=$(OPT) $(WARN)
+CPPFLAGS=-I$(IGBLIB_DIR) -I$(DAEMONS_DIR)/mrpd -I$(MRPCLIENT_DIR) -I$(AVBLIB_DIR) -I$(DAEMONS_DIR)/common
 LDLIBS=-ligb -lpci -lrt -lpthread
-LDFLAGS=-L ../../lib/igb/
+LDFLAGS=-L$(IGBLIB_DIR)
+
+.PHONY: all clean
 
 all: talker listener
 
-talker: talker.o ../../lib/common/avb.o ../common/talker_mrp_client.o
+talker: talker.o $(MRPTALKER_TARGETS) $(AVBLIB_TARGETS)
 
 talker.o: talker.c
-	$(CC) $(CFLAGS) $(INCFLAGS) $(EXTRA_FLAGS) -c talker.c
 
-../common/talker_mrp_client.o: ../common/talker_mrp_client.c ../common/talker_mrp_client.h
-	make -C ../common/ talker_mrp_client.o
-
-listener: listener.o ../../lib/common/avb.o ../common/listener_mrp_client.o
+listener: listener.o $(MRPLISTENER_TARGETS) $(AVBLIB_TARGETS)
 
 listener.o: listener.c
-	$(CC) $(CFLAGS) $(INCFLAGS) $(EXTRA_FLAGS) -c listener.c
 
-../common/listener_mrp_client.o: ../common/listener_mrp_client.c ../common/listener_mrp_client.h
-	make -C ../common/ listener_mrp_client.o
+$(AVBLIB_DIR)/%.o: $(AVBLIB_DIR)/%.h $(AVBLIB_DIR)/%.c
+	make -C $(AVBLIB_DIR) $@
 
-../../lib/common/avb.o: ../../lib/common/avb.c ../../lib/common/avb.h
-	make -C ../../lib/common/ avb.o
+$(MRPCLIENT_DIR)/%.o: $(MRPCLIENT_DIR)/%.c $(MRPCLIENT_DIR)/%.h
+	make -C $(MRPCLIENT_DIR) $@
 
 %: %.o
 	$(CC) $(LDFLAGS) $^ $(LDLIBS) $(EXTRA_FLAGS) -o $@
