+++ examples/simple_listener/Makefile
@@ -1,18 +1,24 @@
+MRPCLIENT_DIR = ../common
+MRPLISTENER_OBJS = listener_mrp_client.o
+MRPLISTENER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPLISTENER_OBJS))
+
+DAEMONS_DIR = ../../daemons
+
 CC ?= gcc
 OPT = -O2 -g
-CFLAGS = $(OPT) -Wall -Wextra -Wno-parentheses
-INCFLAGS = -I../../daemons/mrpd -I../common -I../../daemons/common
+WARN=-Wall -Wextra -Wno-parentheses
+CFLAGS=$(OPT) $(WARN)
+CPPFLAGS = -I$(DAEMONS_DIR)/mrpd -I$(MRPCLIENT_DIR) -I$(DAEMONS_DIR)/common
 LDLIBS = -lpcap -lsndfile -pthread
 
 all: simple_listener
 
-simple_listener: simple_listener.o ../common/listener_mrp_client.o
+simple_listener: simple_listener.o $(MRPLISTENER_TARGETS)
 
 simple_listener.o: simple_listener.c
-	$(CC) $(CFLAGS) $(INCFLAGS) -c simple_listener.c
 
-../common/listener_mrp_client.o: ../common/listener_mrp_client.c ../common/listener_mrp_client.h
-	make -C ../common/ listener_mrp_client.o
+$(MRPCLIENT_DIR)/%.o: $(MRPCLIENT_DIR)/%.c $(MRPCLIENT_DIR)/%.h
+	make -C $(MRPCLIENT_DIR) $@
 
 %: %.o
 	$(CC) $(LDFLAGS) $^ $(LDLIBS) -o $@
