+++ examples/simple_rx/Makefile
@@ -1,23 +1,33 @@
+
+MRPCLIENT_DIR = ../common
+MRPLISTENER_OBJS = listener_mrp_client.o
+MRPLISTENER_TARGETS = $(addprefix $(MRPCLIENT_DIR)/,$(MRPLISTENER_OBJS))
+
+IGBLIB_DIR = ../../lib/igb
+IGBLIB_OBJS = igb.o
+IGBLIB_TARGETS = $(addprefix $(IGBLIB_DIR)/,$(IGBLIB_OBJS))
+
+DAEMONS_DIR = ../../daemons
+
 CC ?= gcc
 OPT = -O2 -g
-CFLAGS = $(OPT) -Wall -Wextra -Wno-parentheses
-IGBDIR = ../../lib/igb
-INCFLAGS = -I../../daemons/mrpd -I../common -I../../daemons/common -I$(IGBDIR)
+WARN=-Wall -Wextra -Wno-parentheses
+CFLAGS=$(OPT) $(WARN)
+CPPFLAGS = -I$(DAEMONS_DIR)/mrpd -I$(MRPCLIENT_DIR) -I$(DAEMONS_DIR)/common -I$(IGBLIB_DIR)
 # GLIBC versions starting with 2.17 don't need -lrt anymore
 LDLIBS = -lpcap -lsndfile -pthread -lpci -lrt
 
 all: simple_rx
 
-simple_rx: simple_rx.o ../common/listener_mrp_client.o $(IGBDIR)/igb.o
+simple_rx: simple_rx.o $(MRPLISTENER_TARGETS) $(IGBLIB_TARGETS)
 
 simple_rx.o: simple_rx.c
-	$(CC) $(CFLAGS) $(INCFLAGS) -c simple_rx.c
 
-../common/listener_mrp_client.o: ../common/listener_mrp_client.c ../common/listener_mrp_client.h
-	make -C ../common/ listener_mrp_client.o
+$(MRPCLIENT_DIR)/%.o: $(MRPCLIENT_DIR)/%.c $(MRPCLIENT_DIR)/%.h
+	make -C $(MRPCLIENT_DIR) $@
 
-$(IGBDIR)/igb.o:  $(IGBDIR)/igb.c
-	make -C $(IGBDIR) igb.o
+$(IGBLIB_DIR)/%.o: $(IGBLIB_DIR)/%.c $(IGBLIB_DIR)/%.h
+	make -C $(IGBLIB_DIR) $@
 
 %: %.o
 	$(CC) $(LDFLAGS) $^ $(LDLIBS) -o $@
