+++ lib/common/avb_srp.h
@@ -1,0 +1,6 @@
+#ifndef __AVB_SRP_H__
+#define __AVB_SRP_H__
+
+#define MRPD_PORT_DEFAULT 7500
+
+#endif
